[feature/scrum-82] 포스트 오타 수정 및 에러날 수 있는 부분 수정

+++ src/main/kotlin/kr/co/wground/post/application/dto/PostSummaryDto.kt
@@ -20,7 +20,7 @@ class PostSummaryDto(
 
 fun List<Post>.toDtos(writers: List<User>, commentsCountById: List<CommentCount>): List<PostSummaryDto> {
     val writerNameByIdMap = writers.associate { it.userId to it.name }
-    val commentsCountByPostId = commentsCountById.associate { id -> id.postId to id.count.toInt() }
+    val commentsCountByPostId = commentsCountById.associate { id -> id.postId to id.count }
 
     return this.map { post ->
         PostSummaryDto(
